Set product to default state and clear cart
The goal of this change was to set each added product to the default
state and clear the cart, when the order was sent.
   First of all, I created a method 'refreshHTML()', which sets the
product to the default state. It finds all options in HTML and
changes them as so as in the object this.data.params. Then sets the
amount widget to 1 using the method AmountWidget.setValue().
   In index.html I needed to give each product an ID to select just the
options and inputs in a single product.
   In Cart.js I decided to not put the new line at the end of
'sendOrder()' into a new method, because it's verry short.