Fix process of emptying cart
After sending any order the cart products were just visually removed.
The array 'thisCart.products' was not emptied. The consequences of
this are these:

 - The number on the top left corner of the cart is not changing.

 - For the future, when you want to validate the order, it will be
   possible to send the order with no ordered products.

The best way to solve this problem is to use the function
'remove()', because it has got a set of operations needed to remove
successfuly a product. The for...of loop doesn't work, because it does
not remove all products (I don't know why).