Validate data of order before sending it
If they are not given any products, a phone number or an address or if
the phone number or the address are incorrect, then it should not be
possible to send the order. The requirements of the both inputs are just
examples.

In case of clicking the submit button, all of the data will be
validated, but if the user changes the value of an input, just this will
be validated. Because these operations are simillar and repeat, I put them
into a new method validate(checkProductsListOrNot, inputs).

It returns also 'true' (when the data is correct) or 'false'. The task
of validate() is not the send the order!

The user will be also informated about each error.

In index.html, the both inputs got new attributes to validate them
automatically. Additionally, the submit button has now 'formnovalidate',
because else there will not be fired the 'submit' event. Instead, I
added 'input.reportValidity()' into validate().